+++ tests/test_interactions.py
@@ -25,6 +25,7 @@ import time
 import json
 
 from selenium.webdriver.common.action_chains import ActionChains
+from selenium.webdriver.common.by import By
 
 
 def create_app(dash_duo):
@@ -44,7 +45,24 @@ def create_app(dash_duo):
         "Node 5": (277, 236),
         "Node 6": (168, 283),
     }
-    return init_pos, actions
+
+    def calculate_edge_position(node_1, node_2):
+        x = round((node_1[0] + node_2[0]) * 0.5)
+        y = round((node_1[1] + node_2[1]) * 0.5)
+        return (x, y)
+
+    edges_positions = {
+        "Edge from Node 1 to Node 2": calculate_edge_position(
+            init_pos["Node 1"], init_pos["Node 2"]
+        ),
+        "Edge from Node 2 to Node 4": calculate_edge_position(
+            init_pos["Node 2"], init_pos["Node 4"]
+        ),
+        "Edge from Node 5 to Node 1": calculate_edge_position(
+            init_pos["Node 5"], init_pos["Node 1"]
+        ),
+    }
+    return init_pos, actions, edges_positions
 
 
 def save_screenshot(dash_duo, dir_name, name):
@@ -73,27 +91,30 @@ def perform_dragging(
     """
     actions.reset_actions()
     actions.move_to_element_with_offset(
-        dash_duo.driver.find_element_by_tag_name("body"), x, y
+        dash_duo.driver.find_element(By.TAG_NAME, "body"), x, y
     )
     actions.drag_and_drop_by_offset(source=None, xoffset=delta_x, yoffset=delta_y)
-    actions.click()
     actions.perform()
     time.sleep(1)
 
     elem_json = json.loads(elem.text)
-    new_pos = elem_json.get("renderedPosition")
-    clicked_label = elem_json.get("data", {}).get("label")
+    new_pos = (
+        elem_json[0].get("position")
+        if "position" in elem_json[0].keys()
+        else elem_json[0].get("renderedPosition")
+    )
+    dragged_label = elem_json[0].get("data", {}).get("label")
 
-    diff_x = round(new_pos["x"] - x)
-    diff_y = round(new_pos["y"] - y)
+    node_x = round(new_pos["x"])
+    node_y = round(new_pos["y"])
 
     save_screenshot(
         dash_duo,
         dir_name,
-        f"Dragged{clicked_label.replace(' ', '')}By{diff_x}x{diff_y}y",
+        f"Dragged{dragged_label.replace(' ', '')}By{node_x}x{node_y}y",
     )
 
-    return diff_x, diff_y
+    return node_x, node_y
 
 
 def perform_clicking(dash_duo, x, y, elem, actions, dir_name="interactions"):
@@ -106,7 +127,7 @@ def perform_clicking(dash_duo, x, y, elem, actions, dir_name="interactions"):
     """
     actions.reset_actions()
     actions.move_to_element_with_offset(
-        dash_duo.driver.find_element_by_tag_name("body"), x, y
+        dash_duo.driver.find_element(By.TAG_NAME, "body"), x, y
     )
     actions.click()
     actions.perform()
@@ -119,54 +140,69 @@ def perform_clicking(dash_duo, x, y, elem, actions, dir_name="interactions"):
     return clicked_label
 
 
-def perform_mouseover(dash_duo, x, y, elem, actions, dir_name="interactions"):
+def perform_mouseover(
+    dash_duo, x, y, elem, actions, dir_name="interactions", screenshot_name="Mouseover"
+):
     actions.reset_actions()
     actions.move_to_element_with_offset(
-        dash_duo.driver.find_element_by_tag_name("body"), x - 50, y
+        dash_duo.driver.find_element(By.TAG_NAME, "body"), x - 50, y
     )
     actions.move_by_offset(50, 0)
     actions.perform()
     time.sleep(1)
 
-    mouseover_label = json.loads(elem.text).get("label")
+    text = json.loads(elem.text)
+    mouseover_label = json.loads(elem.text).get("label") if text else "null"
 
-    save_screenshot(dash_duo, dir_name, "Mouseover" + mouseover_label.replace(" ", ""))
+    save_screenshot(
+        dash_duo, dir_name, screenshot_name + mouseover_label.replace(" ", "")
+    )
 
     return mouseover_label
 
 
 def test_cyin001_dragging(dash_duo):
-    init_pos, actions = create_app(dash_duo)
-
-    drag_error = "Unable to drag Cytoscape nodes properly"
+    init_pos, actions, _ = create_app(dash_duo)
 
     # View module docstring for more information about initial positions
     init_x, init_y = init_pos["Node 1"]
 
+    # Open the Drag data JSON tab
+    actions.move_to_element(dash_duo.find_element("#tabs > div:nth-child(5)"))
+    actions.click().perform()
+    time.sleep(1)
+
     # Select the JSON output element
-    elem_tap = dash_duo.find_element("pre#tap-node-json-output")
+    elem_tap = dash_duo.find_element("pre#elements-data-json-output")
 
-    # Test dragging the nodes around
-    offset_x, offset_y = perform_dragging(
-        dash_duo, init_x, init_y, 0, 0, elem_tap, actions
+    # Get initial positions. Not actualy dragging
+    init_node_x, init_node_y = perform_dragging(
+        dash_duo, init_x, init_y, 1, 1, elem_tap, actions
     )
-    init_x += offset_x
-    init_y += offset_y
 
+    pixels_to_position_conv_factor = 1280 * 0.00085
+    # Test dragging the nodes around
     assert perform_dragging(dash_duo, init_x, init_y, 150, 0, elem_tap, actions) == (
-        150,
-        0,
-    ), drag_error
+        round(150 / pixels_to_position_conv_factor) + init_node_x,
+        0 + init_node_y,
+    )
     assert perform_dragging(
         dash_duo, init_x + 150, init_y, 0, 150, elem_tap, actions
-    ) == (0, 150), drag_error
+    ) == (
+        round(150 / pixels_to_position_conv_factor) + init_node_x,
+        round(150 / pixels_to_position_conv_factor) + init_node_y,
+    )
     assert perform_dragging(
         dash_duo, init_x + 150, init_y + 150, -150, -150, elem_tap, actions
-    ) == (-150, -150), drag_error
+    ) == (init_node_x, init_node_y)
+    assert perform_dragging(dash_duo, init_x, init_y, 100, -100, elem_tap, actions) == (
+        round(100 / pixels_to_position_conv_factor) + init_node_x,
+        round(-100 / pixels_to_position_conv_factor) + init_node_y,
+    )
 
 
 def test_cyin002_clicking(dash_duo):
-    init_pos, actions = create_app(dash_duo)
+    init_pos, actions, _ = create_app(dash_duo)
     click_error = "Unable to click Cytoscape nodes properly"
 
     # Select the JSON output element
@@ -180,8 +216,19 @@ def test_cyin002_clicking(dash_duo):
         ), click_error
 
 
-def test_cyin003_mouseover(dash_duo):
-    init_pos, actions = create_app(dash_duo)
+def test_cyin003_clicking_edges(dash_duo):
+    _, actions, edges_positions = create_app(dash_duo)
+
+    # Select the JSON output element
+    elem_tap = dash_duo.find_element("pre#tap-edge-json-output")
+
+    # Test clicking the edges
+    for label, positions in edges_positions.items():
+        assert perform_clicking(dash_duo, *positions, elem_tap, actions) == label
+
+
+def test_cyin004_mouseover(dash_duo):
+    init_pos, actions, _ = create_app(dash_duo)
     mouseover_error = "Unable to mouseover Cytoscape nodes properly"
 
     # Open the Mouseover JSON tab
@@ -199,3 +246,97 @@ def test_cyin003_mouseover(dash_duo):
             perform_mouseover(dash_duo, *init_pos[label], elem_mouseover, actions)
             == label
         ), mouseover_error
+
+
+def test_cyin005_mouseover_edges(dash_duo):
+    _, actions, edges_positions = create_app(dash_duo)
+
+    # Open the Mouseover JSON tab
+    actions.move_to_element(dash_duo.find_element("#tabs > div:nth-child(3)"))
+    actions.click().perform()
+    time.sleep(1)
+
+    # Select the JSON output element
+    elem_mouseover = dash_duo.find_element("pre#mouseover-edge-data-json-output")
+
+    # Test mouseover the edges
+    for label, positions in edges_positions.items():
+        assert perform_mouseover(dash_duo, *positions, elem_mouseover, actions) == label
+
+
+def test_cyin006_mouseover_unhover(dash_duo):
+    init_pos, actions, _ = create_app(dash_duo)
+    mouseover_error = "Unable to mouseover Cytoscape nodes properly"
+
+    # Open the Mouseover JSON tab
+    actions.move_to_element(dash_duo.find_element("#tabs > div:nth-child(3)"))
+    actions.click().perform()
+    time.sleep(1)
+
+    # Select the JSON output element
+    elem_mouseover = dash_duo.find_element("pre#mouseover-node-data-json-output")
+
+    # Test hovering the nodes
+    label = "Node 1"
+    assert (
+        perform_mouseover(
+            dash_duo,
+            *init_pos[label],
+            elem_mouseover,
+            actions,
+            screenshot_name="MouseoverHover",
+        )
+        == label
+    ), mouseover_error
+
+    # Test unhover
+    label = "null"
+    assert (
+        perform_mouseover(
+            dash_duo,
+            70,
+            250,
+            elem_mouseover,
+            actions,
+            screenshot_name="MouseoverUnhover",
+        )
+        == label
+    ), mouseover_error
+
+
+def test_cyin007_click_twice(dash_duo):
+    init_pos, actions, _ = create_app(dash_duo)
+
+    # Open the Tap Data JSON tab
+    actions.move_to_element(dash_duo.find_element("#tabs > div:nth-child(2)"))
+    actions.click().perform()
+    time.sleep(1)
+
+    # Select the JSON output element
+    elem_tap = dash_duo.find_element("pre#tap-node-data-json-output")
+
+    # Test clicking the same node twice
+    label = "Node 1"
+    x, y = init_pos[label]
+
+    actions.reset_actions()
+    actions.move_to_element_with_offset(
+        dash_duo.driver.find_element(By.TAG_NAME, "body"), x, y
+    )
+    actions.click()
+    actions.perform()
+
+    time.sleep(1)
+    clicked_label_1 = json.loads(elem_tap.text).get("label")
+    clicked_timestamp_1 = json.loads(elem_tap.text).get("timeStamp")
+
+    # Second click
+    actions.click()
+    actions.perform()
+
+    time.sleep(1)
+    clicked_label_2 = json.loads(elem_tap.text).get("label")
+    clicked_timestamp_2 = json.loads(elem_tap.text).get("timeStamp")
+
+    assert clicked_label_1 == clicked_label_2
+    assert clicked_timestamp_1 != clicked_timestamp_2