+++ usage-events.py
@@ -36,7 +36,7 @@ basic_elements = [
             "id": "one-two",
             "source": "one",
             "target": "two",
-            "label": "Edge from Node1 to Node2",
+            "label": "Edge from Node 1 to Node 2",
         }
     },
     {
@@ -116,6 +116,7 @@ app.layout = html.Div(
                     elements=basic_elements,
                     layout={"name": "preset"},
                     style={"height": "500px", "width": "500px"},
+                    clearOnUnhover=True,
                 )
             ],
         ),
@@ -205,6 +206,21 @@ app.layout = html.Div(
                                 )
                             ],
                         ),
+                        dcc.Tab(
+                            label="Drag Data",
+                            children=[
+                                html.Div(
+                                    style=styles["tab"],
+                                    children=[
+                                        html.P("Elements Data JSON:"),
+                                        html.Pre(
+                                            id="elements-data-json-output",
+                                            style=styles["json-output"],
+                                        ),
+                                    ],
+                                )
+                            ],
+                        ),
                     ],
                 ),
             ],
@@ -267,6 +283,14 @@ def displaySelectedNodeData(data):
     Input("cytoscape", "selectedEdgeData"),
 )
 def displaySelectedEdgeData(data):
+    return json.dumps(data, indent=2)
+
+
+@callback(
+    Output("elements-data-json-output", "children"),
+    Input("cytoscape", "elements"),
+)
+def displayElementsData(data):
     return json.dumps(data, indent=2)